refactor(schema): split schema package into cohesive units
Extract schema tree construction, comparison, tag helpers, encoding
compatibility, and statistics decoding from schemanode.go into
separate files to keep sources under the LOC guideline, each with a
matching test file covering typical and edge cases.
