+++ schema/schemanode.go
@@ -3,21 +3,11 @@ package schema
 import (
 	"encoding/json"
 	"fmt"
-	"maps"
-	"runtime"
-	"slices"
 	"strings"
-	"sync"
 
-	"github.com/hangxie/parquet-go/v3/common"
 	"github.com/hangxie/parquet-go/v3/parquet"
-	"github.com/hangxie/parquet-go/v3/reader"
-	pgschema "github.com/hangxie/parquet-go/v3/schema"
-	"github.com/hangxie/parquet-go/v3/types"
-	"golang.org/x/sync/errgroup"
 )
 
-// this represents order of tags in JSON schema and go struct
 var orderedTags = []string{
 	"name",
 	"inname",
@@ -79,670 +69,6 @@ type SchemaOption struct {
 	SkipPageEncoding bool
 }
 
-// readFirstDataPageEncoding reads the first data page header to get its encoding.
-// Uses the parquet-go library's GetFirstDataPageHeader which efficiently reads only
-// the first data page, skipping any dictionary pages.
-func readFirstDataPageEncoding(pr *reader.ParquetReader, rowGroupIndex, columnIndex int) (parquet.Encoding, error) {
-	// Use parquet-go's GetFirstDataPageHeader which correctly handles:
-	// - Dictionary pages at DataPageOffset
-	// - Proper offset calculation including header sizes
-	// - CRC and other page header variations
-	headerInfo, err := pr.GetFirstDataPageHeader(rowGroupIndex, columnIndex)
-	if err != nil {
-		return 0, fmt.Errorf("read first data page header: %w", err)
-	}
-
-	return headerInfo.Encoding, nil
-}
-
-// buildEncodingMap extracts encoding information from row groups by reading the first data page header.
-// For each column, it reads the page header at DataPageOffset to get the actual data page encoding.
-// Note: Parquet files should use consistent encodings across row groups for the same column.
-// This function reads columns in parallel to speed up remote file access.
-func buildEncodingMap(pr *reader.ParquetReader) (map[string]string, error) {
-	result := make(map[string]string)
-
-	// Use the first row group to extract encodings
-	if len(pr.Footer.RowGroups) == 0 {
-		return result, nil
-	}
-
-	columns := pr.Footer.RowGroups[0].Columns
-
-	// Use a mutex to protect concurrent writes to the result map
-	var mu sync.Mutex
-
-	// Process columns in parallel, use runtime.NumCPU() to match available cores
-	g := new(errgroup.Group)
-	g.SetLimit(runtime.NumCPU())
-
-	for colIndex, col := range columns {
-		g.Go(func() error {
-			pathKey := strings.Join(col.MetaData.PathInSchema, common.ParGoPathDelimiter)
-			if col.GetCryptoMetadata() != nil {
-				return nil
-			}
-
-			// Clone the reader to get a dedicated file handle for concurrent access
-			// This is necessary because io.ReadSeeker operations (Seek/Read) are not thread-safe
-			clonedFile, err := pr.PFile.Clone()
-			if err != nil {
-				return fmt.Errorf("failed to clone file for column [%s]: %w", pathKey, err)
-			}
-			defer func() {
-				_ = clonedFile.Close()
-			}()
-
-			// Create a temporary reader with the cloned file
-			clonedReader := &reader.ParquetReader{
-				PFile:         clonedFile,
-				Footer:        pr.Footer,
-				SchemaHandler: pr.SchemaHandler,
-			}
-
-			// Read just the first data page header to get encoding
-			encoding, err := readFirstDataPageEncoding(clonedReader, 0, colIndex)
-			if err != nil {
-				return fmt.Errorf("failed to read encoding for column [%s]: %w", pathKey, err)
-			}
-
-			mu.Lock()
-			result[pathKey] = encoding.String()
-			mu.Unlock()
-			return nil
-		})
-	}
-
-	if err := g.Wait(); err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-// buildCompressionCodecMap extracts compression codec information from the footer metadata.
-// This is a fast operation as it only reads from the already-loaded footer.
-func buildCompressionCodecMap(pr *reader.ParquetReader) map[string]string {
-	result := make(map[string]string)
-
-	if len(pr.Footer.RowGroups) == 0 {
-		return result
-	}
-
-	// Use the first row group to extract compression codecs
-	columns := pr.Footer.RowGroups[0].Columns
-	for _, col := range columns {
-		pathKey := strings.Join(col.MetaData.PathInSchema, common.ParGoPathDelimiter)
-		result[pathKey] = col.MetaData.Codec.String()
-	}
-
-	return result
-}
-
-// bloomFilterInfo holds bloom filter metadata for a column.
-type bloomFilterInfo struct {
-	Enabled bool
-	Size    int32
-}
-
-// buildBloomFilterMap extracts bloom filter information using the correct bitset-only
-// size from SchemaHandler.Infos (populated by the library's detectBloomFilters), rather
-// than raw metadata which includes Thrift header overhead.
-func buildBloomFilterMap(pr *reader.ParquetReader) map[string]bloomFilterInfo {
-	result := make(map[string]bloomFilterInfo)
-
-	if len(pr.Footer.RowGroups) == 0 {
-		return result
-	}
-
-	rootName := pr.SchemaHandler.GetRootInName()
-	columns := pr.Footer.RowGroups[0].Columns
-	for _, col := range columns {
-		if !col.MetaData.IsSetBloomFilterOffset() {
-			continue
-		}
-		pathKey := strings.Join(col.MetaData.PathInSchema, common.ParGoPathDelimiter)
-		fullPath := common.PathToStr(append([]string{rootName}, col.MetaData.GetPathInSchema()...))
-		info := bloomFilterInfo{Enabled: true}
-		if idx, ok := pr.SchemaHandler.MapIndex[fullPath]; ok {
-			info.Size = pr.SchemaHandler.Infos[idx].BloomFilterSize
-		}
-		result[pathKey] = info
-	}
-
-	return result
-}
-
-// BloomFilterSizeMap returns a map from column path (PathInSchema joined) to the correct
-// bloom filter bitset size in bytes, as detected by the parquet-go library from the actual
-// file header (not from metadata which includes Thrift header overhead).
-func BloomFilterSizeMap(pr *reader.ParquetReader) map[string]int32 {
-	result := make(map[string]int32)
-
-	if len(pr.Footer.RowGroups) == 0 {
-		return result
-	}
-
-	rootName := pr.SchemaHandler.GetRootInName()
-	columns := pr.Footer.RowGroups[0].Columns
-	for _, col := range columns {
-		if !col.MetaData.IsSetBloomFilterOffset() {
-			continue
-		}
-		pathKey := strings.Join(col.MetaData.PathInSchema, common.ParGoPathDelimiter)
-		fullPath := common.PathToStr(append([]string{rootName}, col.MetaData.GetPathInSchema()...))
-		if idx, ok := pr.SchemaHandler.MapIndex[fullPath]; ok {
-			result[pathKey] = pr.SchemaHandler.Infos[idx].BloomFilterSize
-		}
-	}
-
-	return result
-}
-
-func NewSchemaTree(reader *reader.ParquetReader, option SchemaOption) (*SchemaNode, error) {
-	// Extract encoding information from the parquet file unless SkipPageEncoding is set
-	var encodingMap map[string]string
-	if !option.SkipPageEncoding {
-		var err error
-		encodingMap, err = buildEncodingMap(reader)
-		if err != nil {
-			return nil, fmt.Errorf("failed to build encoding map: %w", err)
-		}
-	}
-
-	compressionCodecMap := buildCompressionCodecMap(reader)
-
-	// Always extract bloom filter information from footer metadata
-	bloomFilterMap := buildBloomFilterMap(reader)
-
-	schemas := reader.SchemaHandler.SchemaElements
-	root := &SchemaNode{
-		SchemaElement: *schemas[0],
-		Children:      []*SchemaNode{},
-		InNamePath:    []string{schemas[0].Name},
-		ExNamePath:    strings.Split(reader.SchemaHandler.InPathToExPath[schemas[0].Name], common.ParGoPathDelimiter)[:1],
-	}
-	stack := []*SchemaNode{root}
-
-	for pos := 1; len(stack) > 0; {
-		node := stack[len(stack)-1]
-		if option.FailOnInt96 && node.Type != nil && *node.Type == parquet.Type_INT96 {
-			return nil, fmt.Errorf("field [%s] has type INT96 which is not supported", node.Name)
-		}
-		if len(node.Children) < int(node.GetNumChildren()) {
-			childNode := &SchemaNode{
-				SchemaElement: *schemas[pos],
-				Children:      []*SchemaNode{},
-			}
-
-			// append() does not always return new slice, so we need to copy the old slice
-			childNode.InNamePath = make([]string, len(node.InNamePath)+1)
-			copy(childNode.InNamePath, node.InNamePath)
-			childNode.InNamePath[len(node.InNamePath)] = schemas[pos].Name
-
-			inPathKey := strings.Join(childNode.InNamePath, common.ParGoPathDelimiter)
-			childNode.ExNamePath = strings.Split(reader.SchemaHandler.InPathToExPath[inPathKey], common.ParGoPathDelimiter)
-
-			node.Children = append(node.Children, childNode)
-			stack = append(stack, childNode)
-			pos++
-		} else {
-			stack = stack[:len(stack)-1]
-			if len(node.Children) == 0 {
-				node.Children = nil
-			}
-		}
-	}
-
-	populateLeafMetadata(root, encodingMap, compressionCodecMap, bloomFilterMap)
-	markUndefinedSortOrder(root)
-	return root, nil
-}
-
-func populateLeafMetadata(root *SchemaNode, encodingMap, compressionCodecMap map[string]string, bloomFilterMap map[string]bloomFilterInfo) {
-	queue := []*SchemaNode{root}
-	for len(queue) > 0 {
-		node := queue[0]
-		queue = append(queue[1:], node.Children...)
-		node.Name = node.ExNamePath[len(node.ExNamePath)-1]
-
-		if node.Type != nil {
-			pathKey := strings.Join(node.InNamePath[1:], common.ParGoPathDelimiter)
-			if encodingMap != nil {
-				if encoding, found := encodingMap[pathKey]; found {
-					node.Encoding = encoding
-				}
-			}
-			if compressionCodecMap != nil {
-				if codec, found := compressionCodecMap[pathKey]; found {
-					node.CompressionCodec = codec
-				}
-			}
-			if info, found := bloomFilterMap[pathKey]; found && info.Enabled {
-				node.BloomFilter = "true"
-				if info.Size > 0 {
-					node.BloomFilterSize = fmt.Sprint(info.Size)
-				}
-			}
-		}
-	}
-}
-
-// markUndefinedSortOrder recursively marks nodes whose sort order is
-// undefined per the Parquet spec, so that DecodeStatistics skips min/max.
-//   - GEOMETRY, GEOGRAPHY: marked on the node itself (leaf with logical type)
-//   - INTERVAL: marked on the node itself (leaf with converted type)
-//   - VARIANT: a STRUCT whose logical type is on the parent, so all descendants are marked
-func markUndefinedSortOrder(node *SchemaNode) {
-	if node.LogicalType != nil {
-		if node.LogicalType.IsSetGEOMETRY() || node.LogicalType.IsSetGEOGRAPHY() || node.LogicalType.IsSetUNKNOWN() {
-			node.UndefinedSortOrder = true
-			return
-		}
-		if node.LogicalType.IsSetVARIANT() {
-			markAllDescendants(node)
-			return
-		}
-	}
-	if node.ConvertedType != nil && *node.ConvertedType == parquet.ConvertedType_INTERVAL {
-		node.UndefinedSortOrder = true
-		return
-	}
-	for _, child := range node.Children {
-		markUndefinedSortOrder(child)
-	}
-}
-
-func markAllDescendants(node *SchemaNode) {
-	for _, child := range node.Children {
-		child.UndefinedSortOrder = true
-		markAllDescendants(child)
-	}
-}
-
-// CompareOption controls which fields are included in comparison.
-// Zero value compares only the logical schema, ignoring writer directives
-// and root node name.
-type CompareOption struct {
-	CompareEncoding    bool
-	CompareCompression bool
-	CompareBloomFilter bool
-	CompareOmitStats   bool
-	CompareRootName    bool
-}
-
-// ptrEqual returns true if both pointers are nil or both point to equal values.
-func ptrEqual[T comparable](a, b *T) bool {
-	if a == nil && b == nil {
-		return true
-	}
-	if a == nil || b == nil {
-		return false
-	}
-	return *a == *b
-}
-
-// logicalTypeEqual compares two LogicalType pointers using the generated Equals method.
-func logicalTypeEqual(a, b *parquet.LogicalType) bool {
-	if a == nil && b == nil {
-		return true
-	}
-	if a == nil || b == nil {
-		return false
-	}
-	return a.Equals(b)
-}
-
-// logicalSchemaEqual compares the logical schema fields of two SchemaNodes.
-// TypeLength is only compared for FIXED_LEN_BYTE_ARRAY types, and Scale/Precision
-// are only compared for DECIMAL types, since some writers set these to explicit
-// zero pointers for other types while others leave them nil.
-func logicalSchemaEqual(a, b *SchemaNode) bool {
-	if !ptrEqual(a.Type, b.Type) {
-		return false
-	}
-	if a.Type != nil && *a.Type == parquet.Type_FIXED_LEN_BYTE_ARRAY {
-		if !ptrEqual(a.TypeLength, b.TypeLength) {
-			return false
-		}
-	}
-	if !ptrEqual(a.RepetitionType, b.RepetitionType) {
-		return false
-	}
-	if a.Name != b.Name {
-		return false
-	}
-	if !ptrEqual(a.ConvertedType, b.ConvertedType) {
-		return false
-	}
-	if a.ConvertedType != nil && *a.ConvertedType == parquet.ConvertedType_DECIMAL {
-		if !ptrEqual(a.Scale, b.Scale) || !ptrEqual(a.Precision, b.Precision) {
-			return false
-		}
-	}
-	return logicalTypeEqual(a.LogicalType, b.LogicalType)
-}
-
-// writerDirectivesEqual compares writer directive fields based on the given option.
-func writerDirectivesEqual(a, b *SchemaNode, option CompareOption) bool {
-	if option.CompareEncoding && a.Encoding != b.Encoding {
-		return false
-	}
-	if option.CompareCompression && a.CompressionCodec != b.CompressionCodec {
-		return false
-	}
-	if option.CompareBloomFilter && (a.BloomFilter != b.BloomFilter || a.BloomFilterSize != b.BloomFilterSize) {
-		return false
-	}
-	if option.CompareOmitStats && a.OmitStats != b.OmitStats {
-		return false
-	}
-	return true
-}
-
-// IsCompatible performs a structural comparison of two SchemaNode trees.
-// Logical schema fields (Type, TypeLength, RepetitionType, Name, ConvertedType,
-// Scale, Precision, LogicalType, Children) are always compared.
-// The root node name is ignored by default; set CompareRootName to compare it.
-// Writer directive fields (Encoding, CompressionCodec, BloomFilter, BloomFilterSize,
-// OmitStats) are only compared when the corresponding CompareOption flag is set.
-// Derived/metadata fields (InNamePath, ExNamePath, FieldID, NumChildren) are always ignored.
-func (s *SchemaNode) IsCompatible(other *SchemaNode, option CompareOption) bool {
-	if s == nil && other == nil {
-		return true
-	}
-	if s == nil || other == nil {
-		return false
-	}
-
-	if !option.CompareRootName {
-		// Save and temporarily match names so logicalSchemaEqual passes,
-		// then enable name comparison for child nodes.
-		savedName := other.Name
-		other.Name = s.Name
-		defer func() { other.Name = savedName }()
-		option.CompareRootName = true
-	}
-
-	if !logicalSchemaEqual(s, other) {
-		return false
-	}
-	if !writerDirectivesEqual(s, other, option) {
-		return false
-	}
-
-	if len(s.Children) != len(other.Children) {
-		return false
-	}
-	for i := range s.Children {
-		if !s.Children[i].IsCompatible(other.Children[i], option) {
-			return false
-		}
-	}
-
-	return true
-}
-
-func (s *SchemaNode) GetTagMap() map[string]string {
-	tagMap := map[string]string{
-		"repetitiontype": repetitionTypeStr(s.SchemaElement),
-		"type":           typeStr(s.SchemaElement),
-		"name":           s.Name,
-	}
-
-	if len(s.ExNamePath) != 0 {
-		tagMap["name"] = s.ExNamePath[len(s.ExNamePath)-1]
-	}
-
-	if len(s.InNamePath) != 0 {
-		tagMap["inname"] = s.InNamePath[len(s.InNamePath)-1]
-	}
-
-	if tagMap["type"] == "STRUCT" && s.LogicalType == nil && s.ConvertedType == nil {
-		return tagMap
-	}
-
-	if s.Type != nil && *s.Type == parquet.Type_FIXED_LEN_BYTE_ARRAY {
-		tagMap["length"] = fmt.Sprint(*s.TypeLength)
-	}
-
-	if s.LogicalType != nil {
-		s.updateTagFromLogicalType(tagMap)
-	}
-
-	// Add custom parquet-go writer directives before updateTagFromConvertedType
-	// so that LIST/MAP elements can include valueencoding/keyencoding tags
-	if s.Encoding != "" {
-		tagMap["encoding"] = s.Encoding
-	}
-	if s.CompressionCodec != "" {
-		tagMap["compression"] = s.CompressionCodec
-	}
-	if s.OmitStats != "" {
-		tagMap["omitstats"] = s.OmitStats
-	}
-	if s.BloomFilter != "" {
-		tagMap["bloomfilter"] = s.BloomFilter
-	}
-	if s.BloomFilterSize != "" {
-		tagMap["bloomfiltersize"] = s.BloomFilterSize
-	}
-
-	if s.ConvertedType != nil {
-		s.updateTagFromConvertedType(tagMap)
-	}
-
-	return tagMap
-}
-
-func (s *SchemaNode) getTagMapWithPrefix(prefix string) map[string]string {
-	tagMap := s.GetTagMap()
-	ret := map[string]string{}
-	for _, tag := range orderedTags {
-		if tag == "name" || strings.HasPrefix(tag, "key") || strings.HasPrefix(tag, "value") {
-			// these are tags that should never have prefix
-			continue
-		}
-		if val, found := tagMap[tag]; found {
-			ret[prefix+tag] = val
-		}
-	}
-
-	return ret
-}
-
-func (s *SchemaNode) updateTagForList(tagMap map[string]string) {
-	if len(s.Children) == 0 {
-		return
-	}
-
-	if s.Children[0].LogicalType != nil {
-		// LIST => Element (of scalar type)
-		s.Children[0].Name = "Element"
-		*s.Children[0].RepetitionType = parquet.FieldRepetitionType_REQUIRED
-		maps.Copy(tagMap, s.Children[0].getTagMapWithPrefix("value"))
-		return
-	}
-
-	if len(s.Children[0].Children) > 1 {
-		// LIST => Element (of STRUCT)
-		s.Children[0].Name = "Element"
-		s.Children[0].Type = nil
-		s.Children[0].ConvertedType = nil
-		*s.Children[0].RepetitionType = parquet.FieldRepetitionType_REQUIRED
-		return
-	}
-
-	if len(s.Children[0].Children) == 1 {
-		// LIST => List => Element
-		maps.Copy(tagMap, s.Children[0].Children[0].getTagMapWithPrefix("value"))
-		s.Children = s.Children[0].Children
-		s.Children[0].Name = "Element"
-		return
-	}
-}
-
-func (s *SchemaNode) updateTagForMap(tagMap map[string]string) {
-	if len(s.Children) == 0 || s.Children[0] == nil || len(s.Children[0].Children) == 0 {
-		// meaningless interim layer
-		return
-	}
-	if s.Children[0].ConvertedType != nil && *s.Children[0].ConvertedType != parquet.ConvertedType_MAP_KEY_VALUE {
-		// child nodes have been processed
-		return
-	}
-
-	// MAP has schema structure of MAP->MAP_KEY_VALUE->(Field1, Field2)
-	// expected output is MAP->(Key, Value)
-	maps.Copy(tagMap, s.Children[0].Children[0].getTagMapWithPrefix("key"))
-	maps.Copy(tagMap, s.Children[0].Children[1].getTagMapWithPrefix("value"))
-	s.Children = s.Children[0].Children[0:2]
-	s.Children[0].Name = "Key"
-	s.Children[1].Name = "Value"
-}
-
-func (s *SchemaNode) updateTagFromConvertedType(tagMap map[string]string) {
-	if s.ConvertedType == nil {
-		return
-	}
-
-	tagMap["convertedtype"] = s.ConvertedType.String()
-
-	switch *s.ConvertedType {
-	case parquet.ConvertedType_LIST:
-		s.updateTagForList(tagMap)
-	case parquet.ConvertedType_MAP:
-		s.updateTagForMap(tagMap)
-	case parquet.ConvertedType_DECIMAL:
-		tagMap["scale"] = fmt.Sprint(*s.Scale)
-		tagMap["precision"] = fmt.Sprint(*s.Precision)
-		if *s.Type == parquet.Type_FIXED_LEN_BYTE_ARRAY {
-			tagMap["length"] = fmt.Sprint(*s.TypeLength)
-		}
-	case parquet.ConvertedType_INTERVAL:
-		tagMap["length"] = "12"
-	}
-}
-
-func (s *SchemaNode) updateTagFromLogicalType(tagMap map[string]string) {
-	if s.LogicalType == nil {
-		return
-	}
-
-	switch {
-	case s.LogicalType.IsSetBSON():
-		tagMap["logicaltype"] = "BSON"
-	case s.LogicalType.IsSetDATE():
-		tagMap["logicaltype"] = "DATE"
-	case s.LogicalType.IsSetDECIMAL():
-		tagMap["logicaltype"] = "DECIMAL"
-		tagMap["logicaltype.precision"] = fmt.Sprint(s.LogicalType.DECIMAL.Precision)
-		tagMap["logicaltype.scale"] = fmt.Sprint(s.LogicalType.DECIMAL.Scale)
-	case s.LogicalType.IsSetENUM():
-		tagMap["logicaltype"] = "ENUM"
-	case s.LogicalType.IsSetFLOAT16():
-		tagMap["logicaltype"] = "FLOAT16"
-	case s.LogicalType.IsSetGEOGRAPHY():
-		tagMap["logicaltype"] = "GEOGRAPHY"
-	case s.LogicalType.IsSetGEOMETRY():
-		tagMap["logicaltype"] = "GEOMETRY"
-	case s.LogicalType.IsSetINTEGER():
-		tagMap["logicaltype"] = "INTEGER"
-		tagMap["logicaltype.bitwidth"] = fmt.Sprint(s.LogicalType.INTEGER.BitWidth)
-		tagMap["logicaltype.issigned"] = fmt.Sprint(s.LogicalType.INTEGER.IsSigned)
-	case s.LogicalType.IsSetJSON():
-		tagMap["logicaltype"] = "JSON"
-	case s.LogicalType.IsSetVARIANT():
-		// VARIANT is a semi-structured logical type introduced in newer parquet-format
-		tagMap["logicaltype"] = "VARIANT"
-		for _, child := range s.Children {
-			if child.Encoding != "" {
-				tagMap["encoding"] = child.Encoding
-			}
-			if child.CompressionCodec != "" {
-				tagMap["compression"] = child.CompressionCodec
-			}
-			if tagMap["encoding"] != "" && tagMap["compression"] != "" {
-				break
-			}
-		}
-	case s.LogicalType.IsSetSTRING():
-		tagMap["logicaltype"] = "STRING"
-	case s.LogicalType.IsSetTIME():
-		tagMap["logicaltype"] = "TIME"
-		tagMap["logicaltype.isadjustedtoutc"] = fmt.Sprint(s.LogicalType.TIME.IsAdjustedToUTC)
-		tagMap["logicaltype.unit"] = timeUnitToTag(s.LogicalType.TIME.Unit)
-	case s.LogicalType.IsSetTIMESTAMP():
-		tagMap["logicaltype"] = "TIMESTAMP"
-		tagMap["logicaltype.isadjustedtoutc"] = fmt.Sprint(s.LogicalType.TIMESTAMP.IsAdjustedToUTC)
-		tagMap["logicaltype.unit"] = timeUnitToTag(s.LogicalType.TIMESTAMP.Unit)
-	case s.LogicalType.IsSetUUID():
-		tagMap["logicaltype"] = "UUID"
-	case s.LogicalType.IsSetUNKNOWN():
-		tagMap["logicaltype"] = "UNKNOWN"
-	}
-}
-
-func (s *SchemaNode) GetPathMap() map[string]*SchemaNode {
-	retVal := map[string]*SchemaNode{}
-	queue := []*SchemaNode{s}
-	for len(queue) > 0 {
-		node := queue[0]
-		queue = append(queue[1:], node.Children...)
-		retVal[strings.Join(node.InNamePath[1:], common.ParGoPathDelimiter)] = node
-	}
-	return retVal
-}
-
-// GetInExNameMap returns a map from internal name path to external name path
-// for all nodes in the schema tree via BFS traversal.
-func (s *SchemaNode) GetInExNameMap() map[string][]string {
-	retVal := map[string][]string{}
-	queue := []*SchemaNode{s}
-	for len(queue) > 0 {
-		node := queue[0]
-		queue = append(queue[1:], node.Children...)
-		inPath := strings.Join(node.InNamePath[1:], common.ParGoPathDelimiter)
-		retVal[inPath] = node.ExNamePath[1:]
-	}
-	return retVal
-}
-
-func typeStr(se parquet.SchemaElement) string {
-	if se.Type != nil {
-		return se.Type.String()
-	}
-	if se.LogicalType != nil && se.LogicalType.IsSetVARIANT() {
-		return "VARIANT"
-	}
-	if se.ConvertedType != nil {
-		return se.ConvertedType.String()
-	}
-	return "STRUCT"
-}
-
-func repetitionTypeStr(se parquet.SchemaElement) string {
-	if se.RepetitionType == nil {
-		return "REQUIRED"
-	}
-	return se.RepetitionType.String()
-}
-
-func timeUnitToTag(timeUnit *parquet.TimeUnit) string {
-	if timeUnit == nil {
-		return ""
-	}
-	if timeUnit.IsSetNANOS() {
-		return "NANOS"
-	}
-	if timeUnit.IsSetMICROS() {
-		return "MICROS"
-	}
-	if timeUnit.IsSetMILLIS() {
-		return "MILLIS"
-	}
-	return "UNKNOWN_UNIT"
-}
-
 func (s SchemaNode) GoStruct(forceCamelCase bool) (string, error) {
 	goStruct, err := goStructNode{
 		SchemaNode:     s,
@@ -792,124 +118,4 @@ func removeTagFromString(tagString, tagName string) string {
 		}
 	}
 	return strings.Join(filtered, ", ")
-}
-
-// EncodingToString converts a slice of parquet encodings to sorted strings.
-func EncodingToString(encodings []parquet.Encoding) []string {
-	ret := make([]string, len(encodings))
-	for i := range encodings {
-		ret[i] = encodings[i].String()
-	}
-	slices.Sort(ret)
-	return ret
-}
-
-// encodingCompatibilityMap maps Parquet data types to their compatible encodings
-// (excluding PLAIN, which is universally compatible).
-//
-// Per Parquet spec and parquet-go validation rules:
-//   - RLE: BOOLEAN, INT32, INT64 only
-//   - DELTA_BINARY_PACKED: INT32, INT64 only
-//   - DELTA_BYTE_ARRAY, DELTA_LENGTH_BYTE_ARRAY: BYTE_ARRAY only
-//   - BYTE_STREAM_SPLIT: FLOAT, DOUBLE, INT32, INT64, FIXED_LEN_BYTE_ARRAY
-//   - PLAIN_DICTIONARY: All types (v1 data pages only, validated separately)
-var encodingCompatibilityMap = map[string][]string{
-	"BOOLEAN":              {"BIT_PACKED", "PLAIN_DICTIONARY", "RLE", "RLE_DICTIONARY"},
-	"BYTE_ARRAY":           {"DELTA_BYTE_ARRAY", "DELTA_LENGTH_BYTE_ARRAY", "PLAIN_DICTIONARY", "RLE_DICTIONARY"},
-	"DOUBLE":               {"BYTE_STREAM_SPLIT", "PLAIN_DICTIONARY", "RLE_DICTIONARY"},
-	"FIXED_LEN_BYTE_ARRAY": {"BYTE_STREAM_SPLIT", "DELTA_BYTE_ARRAY", "PLAIN_DICTIONARY", "RLE_DICTIONARY"},
-	"FLOAT":                {"BYTE_STREAM_SPLIT", "PLAIN_DICTIONARY", "RLE_DICTIONARY"},
-	"INT32":                {"BYTE_STREAM_SPLIT", "DELTA_BINARY_PACKED", "PLAIN_DICTIONARY", "RLE_DICTIONARY"},
-	"INT64":                {"BYTE_STREAM_SPLIT", "DELTA_BINARY_PACKED", "PLAIN_DICTIONARY", "RLE_DICTIONARY"},
-}
-
-// GetAllowedEncodings returns all encodings compatible with the given Parquet data type.
-func GetAllowedEncodings(dataType string) []string {
-	dataType = strings.ToUpper(dataType)
-
-	allowed, exists := encodingCompatibilityMap[dataType]
-	if !exists {
-		return []string{"PLAIN"}
-	}
-
-	return append([]string{"PLAIN"}, allowed...)
-}
-
-// ConvertedTypeString returns the formatted converted type string from schema tags.
-// Returns "" if node is nil or has no converted type information.
-func (node *SchemaNode) ConvertedTypeString() string {
-	if node == nil {
-		return ""
-	}
-
-	tagMap := node.GetTagMap()
-
-	var parts []string
-	for _, tag := range orderedTags {
-		if tag != "convertedtype" && tag != "scale" && tag != "precision" && tag != "length" {
-			continue
-		}
-		if value, found := tagMap[tag]; found {
-			parts = append(parts, tag+"="+value)
-		}
-	}
-
-	return strings.Join(parts, ", ")
-}
-
-// LogicalTypeString returns the formatted logical type string from schema tags.
-// Returns "" if node is nil or has no logical type information.
-func (node *SchemaNode) LogicalTypeString() string {
-	if node == nil {
-		return ""
-	}
-
-	tagMap := node.GetTagMap()
-
-	var parts []string
-	for _, tag := range orderedTags {
-		if !strings.HasPrefix(tag, "logicaltype") {
-			continue
-		}
-		if value, found := tagMap[tag]; found {
-			parts = append(parts, tag+"="+value)
-		}
-	}
-
-	return strings.Join(parts, ", ")
-}
-
-// IsEncodingCompatible reports whether the given encoding is compatible with the data type.
-func IsEncodingCompatible(encoding, dataType string) bool {
-	if dataType == "" {
-		return false
-	}
-
-	encoding = strings.ToUpper(encoding)
-	dataType = strings.ToUpper(dataType)
-
-	if encoding == "PLAIN" {
-		return true
-	}
-
-	compatibleEncodings, exists := encodingCompatibilityMap[dataType]
-	if !exists {
-		return false
-	}
-
-	return slices.Contains(compatibleEncodings, encoding)
-}
-
-// DecodeStatistics decodes a parquet.Statistics object into human-readable min/max values.
-// Returns (nil, nil) for nil/empty statistics, nodes with undefined sort order, or on error.
-func (node *SchemaNode) DecodeStatistics(stats *parquet.Statistics) (min, max any) {
-	if node.UndefinedSortOrder {
-		return nil, nil
-	}
-	rawMin, rawMax, err := pgschema.DecodeStatisticsMinMax(&node.SchemaElement, stats)
-	if err != nil {
-		return nil, nil
-	}
-	return types.ConvertToJSONType(rawMin, &node.SchemaElement),
-		types.ConvertToJSONType(rawMax, &node.SchemaElement)
 }
